further tuning

+++ src/levbv.h
@@ -22,20 +22,22 @@ extern "C" {
 
 typedef struct {
     void **keys;
-    uint32_t *lens;
-    int capacity;
-    int elements;
+    size_t *lens;
+    size_t capacity;
+    size_t elements;
 } Array;
 
-int dist_bytes (const unsigned char * a, int alen, unsigned const char * b,  int blen);
-int dist_utf8_ucs (unsigned char * a, uint32_t alen, unsigned char * b, uint32_t blen);
-int dist_uni (const uint32_t *a, int alen, const uint32_t *b, int blen);
-int dist_hybrid (const uint32_t *a, int alen, const uint32_t *b, int blen);
-int dist_simple (const uint32_t *a, int alen, const uint32_t *b, int blen);
-int dist_simple_utf8 (unsigned char * a, uint32_t alen, unsigned char * b, uint32_t blen);
+int dist_bytes    (const unsigned char * a, int alen, unsigned const char * b, int blen);
+int dist_utf8_ucs (unsigned char * a, size_t alen, unsigned char * b, size_t blen);
+int dist_uni      (const uint32_t *a, size_t alen, const uint32_t *b, size_t blen);
+int dist_hybrid   (const uint32_t *a, size_t alen, const uint32_t *b, size_t blen);
+int dist_simple   (const uint32_t *a, size_t alen, const uint32_t *b, size_t blen);
+int dist_simple_utf8 (unsigned char * a, size_t alen, unsigned char * b, size_t blen);
 
-int dist_array (const Array *a, const Array *b);
+int dist_array      ( const Array *a, const Array *b );
 int dist_simple_arr ( const Array *a, const Array *b );
+
+//void *stos_memset(void *s, int c, size_t n);
 
 #endif
